tests/wpt/web-platform-tests/css/work-in-progress/gadams/cssom/csspagerule-interface.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD//XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
  <head>
    <title>CSSOM - CSSPageRule - Interface</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="http://www.w3.org/TR/cssom-1/#the-csspagerule-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"/>
    <script src="/resources/testharnessreport.js"/>
  </head>
  <body>
    <div id="log"></div>
    <div id="box"></div>
    <script><![CDATA[

      test(function(){

        assert_true('CSSPageRule' in window);

      }, '\'CSSPageRule\' in window');

      test(function(){

        assert_true(CSSPageRule instanceof Function);

      }, 'CSSPageRule instanceof Function');

      test(function(){

        assert_true('prototype' in CSSPageRule);

      }, '\'prototype\' in CSSPageRule');

      test(function(){

        assert_equals(CSSPageRule.prototype.toString(), '[object CSSPageRulePrototype]');

      }, 'CSSPageRule.prototype.toString() == \'[object CSSPageRulePrototype]\'');

      test(function(){

        assert_true('constructor' in CSSPageRule.prototype);

      }, '\'constructor\' in CSSPageRule.prototype');

      test(function(){

        assert_equals(CSSPageRule.prototype.constructor,CSSPageRule);

      }, 'CSSPageRule.prototype.constructor === CSSPageRule');

    ]]>
    </script>
  </body>
</html>
